[Authorize.net] Various corrections to L2/L3 data inputs (#265)
Our L2/L3 Line Item inputs were not fully compliant with Authorize.net's specification.

Firstly, we were sending the incorrect item fields. CommodityCode was used for the item ID when instead ProductCode should be used. And ProductCode was used for the name when Description should be used.

Next, Authorize.net requires an item ID and a name on every item. Some cart items may not have a product code or description added. To facilitate successful auths when limited data is available, we will use the item index as a substitute for item ID, and the item ID as substitute for the item name. There is precedent for adding this logic for other L2/L3-supported processors.

Lastly, I renamed some types and fields and added notes regarding the specific requirements for each field from Authorize.net's documentation.